api/requirements.txt
fastapi
uvicorn
websockets
python-lorem
